% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionicons.R
\name{as_array}
\alias{as_array}
\title{'ionicons' png file}
\usage{
as_array(name, width = NULL, height = NULL, fill = "black")
}
\arguments{
\item{name}{icon name (without extensions)}

\item{width}{output width in pixels or NULL for default.}

\item{height}{output height in pixels or NULL for default.}

\item{fill}{fill color}
}
\description{
raster of an icon from 'ionicons' pack.
Width, height and fill color can be specified.
}
\details{
All icons are 512 pixels * 512 pixels by default.
}
\examples{
as_array(name = "alert")
}

