% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionicons.R
\name{ionicons}
\alias{ionicons}
\title{'ionicons' png file}
\usage{
ionicons(name)
}
\arguments{
\item{name}{icon name (without extensions)}
}
\description{
get full path of an icon from 'ionicons' pack.
}
\details{
All icons are squares (width 512 pixels and height 512 pixels).
}
\examples{
ionicons(name = "alert")
}

