% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionicons.R
\name{ls_ionicons}
\alias{ls_ionicons}
\title{list 'ionicons' icons}
\usage{
ls_ionicons(pattern = NULL)
}
\arguments{
\item{pattern}{regexp pattern to use to find icons names}
}
\description{
get 'ionicons' pack icons names.
}
\examples{
ls_ionicons()
}

